requirements.txt
flask~=3.0.0
google-cloud-language
google-cloud-aiplatform
google-cloud-speech
google-cloud-documentai
google-cloud-storage
pandas
pydub